h update

+++ 2025.09.27-Homework-1/t10.cpp
@@ -3,8 +3,8 @@
 int main(int argc, char** argv)
 {
 	long long a=0;
-	int b=0;
-	int c=0;
+	long long b=0;
+	long long c=0;
 	scanf("%lld %lld %lld", &a, &b, &c);
     printf("%lld", a+b+c);
 	return 0;
